[FIX] account: general ledger, currency displayed twice
When printing the general ledger with "With Currency"
checked, two currencies were displayed in the colum
"Currency" when an amount was set, one time the company
currency, one time the move line currency. Only
the move line currency should be displayed.

Besides, the total of currencies amount
for each account should not display the currency symbol
at all, as the total may be composed of multiple currencies.

opw-632086