[FIX] css: ie9 specific fixes
Create a new stylesheet for IE9 and less custom css.

The issue #6198 and #6201 is present in IE9 since a div with height 100% inside
a table-cell (i.e: td tag) element only take the minimum height possible for the
content.

For issue #6198: in IE9 a kanban tile could not be moved to an empty kanban,
this fix sets a minimum height for a empty column (80px, the size of two empty
tiles) so now a tile can be moved in the 80px in the top of the column.

For issue #6201: graphs took the minimum height in IE9, this fix set a fixed
height to improve the situation.

For issue #6200: in IE9, the list view could wrongly not go the next line
even if its width is 100%. For the given case, the display CSS property of
these listview is set to value "block".

opw-632415 opw-632417 opw-632416