[FIX] point_of_sale: Taxes in sales lines report
product_taxes_rel is the many2many table between
product.template and account.tax

product_id on the POS order line is a product.product.

Therefore, the join on product_taxes_rel should
be done using the product.template id of
the product on the pos order line, and
not directly the product.product id.

opw-632720