[IMP] crm: Opportunities Analysis menu visible to Sale Manager
Before this commit, it was only visible by the admin user.

I used the crm.phonecall.report rules as an example, but I don't
think either of those or the opportunities ones are actually
used in the case of group_sale_salesman, because the submenu
Sales of the Reporting menu is only visible to the Sale Manager.

Fixes opw 615048.