[FIX] account: float issue in reconciliation wizard
The CSS on the linei with the "New" button of an bank statement
reconciliation wizard was only applied when there was only a multiple of
two "block" before.

Hence it worked when there was a multiple of 2 cases before, but if not
(for example when account_analytic_plans is installed) on:

- google chrome: the button is on the right (instead of on a line by its
  own).

- firefox: idem, but also a difference in float display positioning
  caused the parent container to not take floating elements into account
  to calculate it's height.

This fix apply the CSS independently if the last block is even or odd.

Before this fix in firefox:
 https://cloud.githubusercontent.com/assets/9977887/7315195/70ca5eca-ea6b-11e4-8892-1272f7ee0cb4.png

After this fix in firefox:
 https://cloud.githubusercontent.com/assets/9977887/7315194/6e2b05ca-ea6b-11e4-9400-69c9cd587756.png

closes #6459
opw-633703