[FIX] fields: do not infer a related field to be required
Forcing a related field to be required causes some issues.  Having such a field
in a one2many relation may block the creation of records in the field, as the
main record does not exist yet, and the field has no value.

It may also cause trouble if the field is stored: being required constrains the
corresponding column to be NOT NULL.  Inserting a row will fail, because
related fields are currently determined *after* the row insertion.