[FIX] website_sale: display real price unit.
Now we display the price_reduce from the sale order line
and not the unit price.

If a discount has been added to the quotation, the user had
in the front end not the real price for the line, but the
discounted total which one is the total of the sale order.

So we need to display the final unit price (= reduce_price) to
be consistent with the total price displayed.