[FIX] web: default loaded view on direct access/refresh
When loading an action using a specific url, with
a view_type different than the default view of the action
(e.g. Project > Project > Issues has as view order
kanban, list, calendar, form, graph, and try to load
this menu action from an url, but by specifying another view_type
than 'kanban', for instance, 'list'), the default loaded view was
always the first view, and then a switch was done to go to the view_list.

It caused issues when using a filter needing a field available in the list
view but not in the kanban view.

It also lead to an issue in the reporting, when trying to load a reporting
action with a view not being the first one of the action
(e.g. Reporting > Project > Issues Analysis, switch to the second view
'Graph', and then refresh -> Error).

This revision makes so the view from the view_type url arg is directly loaded.

opw-640289