[IMP] account: performances on total_invoiced field
The computation of total_invoiced field was very slow when the size of
account.invoice.report grows. This is due to usage of the field
user_currency_price_total that requires to build the full view (generates query
(id in []) for function field).

Using temporary SQL view (inspired by caf333e), directly filter the needed
items and avoid building the full table.

Fixes #6654