[FIX] hr_gamification: user_id not required
Revert the modification from commit 8ae67f6a5412ccde126bf8c85849f33f387df198 ; it appears that the test of required field is done before the computation of the related field and that leads to a Validation Error.

As the field employee_id is already required and the user_id is related to this field, the attribute required is redundant and unecessary.