[IMP] base: remove multi_company.default
Multi_company.default was used to set the default company on newly created
records in a multi-company environment. We remove it because it is redundant
with User-Defined Default.

The function _company_default_get is left for backward compatibility and to
allow the creation of modules that implement more complex policies.